chore: refactor clerk internals to be testable
