+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const NO_NOTES_BODY = '**No Release Notes**';
+export const NOTES_LEAD = '**Release Notes Persisted**';
