+++ src/index.ts
@@ -1,45 +1,13 @@
 import { Application, Context } from 'probot';
 import { WebhookPayloadWithRepository } from 'probot/lib/context';
-
-const OMIT_FROM_RELEASE_NOTES_KEYS = [
-  'no-notes',
-  'no notes',
-  'no_notes',
-  'none',
-  'no',
-  'nothing',
-  'empty',
-  'blank',
-];
-
-const getReleaseNotes = (pr: WebhookPayloadWithRepository['pull_request']) => {
-  const currentPRBody = pr.body;
-
-  const onelineMatch = /(?:(?:\r?\n)|^)notes: (.+?)(?:(?:\r?\n)|$)/gi.exec(currentPRBody);
-  const multilineMatch =
-      /(?:(?:\r?\n)Notes:(?:\r?\n)((?:\*.+(?:(?:\r?\n)|$))+))/gi.exec(currentPRBody);
-
-  let notes: string | null = null;
-  if (onelineMatch && onelineMatch[1]) {
-    notes = onelineMatch[1];
-  } else if (multilineMatch && multilineMatch[1]) {
-    notes = multilineMatch[1];
-  }
-
-  // remove the default PR template
-  if (notes) {
-    notes = notes.replace(/<!--.*?-->/g, '');
-  }
-
-  return notes;
-};
+import * as noteUtils from './note-utils';
 
 const submitFeedbackForPR = async (
   context: Context,
   pr: WebhookPayloadWithRepository['pull_request'],
   shouldComment = false,
 ) => {
-  const releaseNotes = getReleaseNotes(context.payload.pull_request);
+  const releaseNotes = noteUtils.findNoteInPRBody(context.payload.pull_request.body);
   if (!releaseNotes) {
     await context.github.repos.createStatus(context.repo({
       state: 'failure' as 'failure',
@@ -58,20 +26,8 @@ const submitFeedbackForPR = async (
   }
 
   if (shouldComment) {
-    const NO_NOTES_BODY = '**No Release Notes**';
-    const NOTES_LEAD = '**Release Notes Persisted**';
-
-    let body = NO_NOTES_BODY;
-    if (releaseNotes && (OMIT_FROM_RELEASE_NOTES_KEYS.indexOf(releaseNotes) === -1)) {
-      const splitNotes = releaseNotes.split('\n').filter(line => line !== '');
-      if (splitNotes.length > 0) {
-        const notes = splitNotes.map(line => `> ${line}`).join('\n');
-        body = `${NOTES_LEAD}\n\n${notes}`;
-      }
-    }
-
     await context.github.issues.createComment(context.repo({
-      body,
+      body: noteUtils.createPRCommentFromNotes(releaseNotes),
       number: pr.number,
     }));
   }
